feat(dashboard): event-driven refresh, fix hang, OTel observability (v0.4.4)
- Replace 10s polling with PostgreSQL LISTEN/NOTIFY trigger on memories table
- Fix dashboard hang: replace localhost with 127.0.0.1 (DNS resolution blocked
  after customtkinter + OTel init on Windows)
- Dashboard uses JSONL-only span exporter (no OTLP gRPC, no ConsoleSpanExporter)
- fix: fetch_obs_metrics reads otel-traces.jsonl instead of legacy open-brain.jsonl
- fix: check_mcp uses OTel trace recency instead of WSL pgrep
- fix: OTLP exporter timeout in telemetry.py (2s timeout, 3s export deadline)
- Set Windows AppUserModelID + DPI awareness for crisp taskbar icon
- Add logs/ to .gitignore, remove tracked log files
- Update README dashboard section, CHANGELOG v0.4.4, demo slide deck

+++ server.py
@@ -566,8 +566,8 @@ def db_search(
             SELECT v.id, v.content, v.metadata, v.created_at, v.project,
                    v.annotation, v.upvotes, v.downvotes, v.access_count,
                    round(
-                       ({vec_weight} * v.vec_score + {fts_weight} * COALESCE(f.fts_score, 0))
-                       {decay_expr}
+                       (({vec_weight} * v.vec_score + {fts_weight} * COALESCE(f.fts_score, 0))
+                       {decay_expr})::numeric
                    , 4) AS similarity
             FROM vec_q v
             LEFT JOIN fts_q f ON v.id = f.id
